% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{allXYZ}
\alias{allXYZ}
\title{data/allXYZ　調査区ランドマーク座標　(Nakajima)}
\format{
An object of class `data.frame`.
}
\usage{
data(allXYZ)
}
\description{
data/allXYZ　調査区ランドマーク座標　(Nakajima)
}
\examples{
allXYZ

}
\keyword{datasets}
