% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{CalculationInfo}
\alias{CalculationInfo}
\title{集計計算に用いるファイル名・スプレッドシートの列名　調査年と調査区名から返します
VegetationSurveyYearsに基づきます
各調査期の列名が正しく入力されていることを確認してください。}
\usage{
CalculationInfo(year = 2025, plot_name_jp = "有峰")
}
\arguments{
\item{plot_name_jp}{}
}
\description{
集計計算に用いるファイル名・スプレッドシートの列名　調査年と調査区名から返します
VegetationSurveyYearsに基づきます
各調査期の列名が正しく入力されていることを確認してください。
}
\examples{
VegetationSurveyYears
#各調査期の列名が正しく入力されていることを確認してください!!!

CalculationInfo(2025,"有峰")

#yachou  :エクセル等で入力したオリジナルの野帳ファイル名です
#         保存する際はこの名前と同じにしてください【.xlsxがこのルールで命名されています。
 .csvとして保存すると自動的にこの名前で保存されます】
#yachou2 :オリジナルを整理した野帳ファイル名です
#yachou :植生被度の列名となります

# 計算用データ・フレーム行列名対応表　yachou_calc.colnames

}
\seealso{
\code{\link{SpeciesNameCheck}}, \code{\link{SpeciesNameCorrect}}, \code{\link{FieldNote_CheckCorrect}}
}
