% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegetationSurvey.R
\name{Domin_Krajina}
\alias{Domin_Krajina}
\title{クロステーブル作成用　被度→被度階級(Domin_Krajina)}
\usage{
Domin_Krajina(cv)
}
\arguments{
\item{cv}{}
}
\description{
クロステーブル作成用　被度→被度階級(Domin_Krajina)
}
\examples{

Domin_Krajina(26.5)

cover<-seq(0,100,1)
plot(cover, Domin_Krajina(cover),type="l")


}
