% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{DT_VCrepo}
\alias{DT_VCrepo}
\title{報告書用植生調査集計表　Data Table of Vegetation Chronosequence}
\usage{
DT_VCrepo(plot_name = "Mimatsu")
}
\arguments{
\item{plot_name}{}
}
\description{
報告書用植生調査集計表　Data Table of Vegetation Chronosequence
}
\examples{

DT_VCrepo("Mimatsu")
DT_VCrepo("Arimine")

# for (i in plt$plot_name) {DT_VCrepo(i)}

}
