% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{EncodingExchange}
\alias{EncodingExchange}
\title{csvファイルのエンコーディング変換}
\usage{
EncodingExchange(filename = "test.csv", en_old = "cp932", en_new = "utf-8")
}
\arguments{
\item{en_old}{original fileEncoding　defaulut　"cp932"}

\item{en_new}{new  fileEncoding  defaulut　"utf-8"}
}
\description{
上書きします。必要に応じて事前にバックアップしておいてください。
}
\examples{

#EncodingExchange(filename="H30植生調査野帳_美松.csv",en_old="cp932",en_new="utf-8")

　#EncodingExchange(filename="H31植生調査野帳_有峰.csv",en_old="cp932",en_new="utf-8")
　
　
　
}
