% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{FieldNote_CheckCorrect}
\alias{FieldNote_CheckCorrect}
\title{野帳の整理をします。
生野帳データの階層クラスの確認、修正、未入力データの補完をします
　サブプロット、階層、種名の順でソートします。
　サブプロットに同一種が複数ある場合は被度を合算し統一します}
\usage{
FieldNote_CheckCorrect(d = FieldNote)
}
\arguments{
\item{KAISO_colnm}{}
}
\description{
野帳の整理をします。
生野帳データの階層クラスの確認、修正、未入力データの補完をします
　サブプロット、階層、種名の順でソートします。
　サブプロットに同一種が複数ある場合は被度を合算し統一します
}
\examples{
d<-FieldNote_Arimine2019_raw
FieldNote_CheckCorrect(d)



}
