% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{Fig_DominanceValue_layer}
\alias{Fig_DominanceValue_layer}
\title{Title}
\usage{
Fig_DominanceValue_layer(
  plot_name = "Bijodaira",
  period = 5,
  vc_repo = VCrepo,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

Fig_DominanceValue_layer("Arimine")
Fig_DominanceValue_layer(plot_name ="Joudosan")
Fig_DominanceValue_layer(plot_name ="Midagahara")

}
