% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{flora_new}
\alias{flora_new}
\title{植物目録floraの更新
flora　→　flora2　外来種追加の事例}
\usage{
flora_new(sp = c("アイバソウ", "マルバヤナギ", "アライトツメクサ"))
}
\arguments{
\item{sp}{}
}
\description{
植物目録floraの更新
flora　→　flora2　外来種追加の事例
}
\examples{

(d<-flora_new(NonNativePlants_add2025))
d$form[match(c("マルバヤナギ","ドロノキ"),d$spj])<-"bl"
d$form[match(c("ヤマハギ"),d$spj)]<-"bs"
d

flora2<-rbind(flora,d)

# usethis::use_data(flora2,overwrite=T)

}
