% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{forest_code}
\alias{forest_code}
\title{data/ 植生モニタリング調査(石田仁・中島春樹)}
\format{
An object of class `tbl_df`.
}
\source{
@examples
forest_code
 SpeciesNameCheck(forest_code$sp)
}
\usage{
data(vv5)
}
\description{
data/ 植生モニタリング調査(石田仁・中島春樹)
}
\keyword{datasets}
