% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{ForestFloraCode}
\alias{ForestFloraCode}
\title{森林調査での植物目録の各調査区における胸高断面積割合コードの作成
9-10はAに変換}
\usage{
ForestFloraCode(cod = m[["オオシラビソ"]])
}
\arguments{
\item{cod}{}
}
\description{
森林調査での植物目録の各調査区における胸高断面積割合コードの作成
9-10はAに変換
}
\examples{

ForestFloraCode(1:10)

}
