% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{ForestFloraMatrix}
\alias{ForestFloraMatrix}
\title{森林調査での各調査区-出現全種の行列表}
\usage{
ForestFloraMatrix(d = dd5)
}
\arguments{
\item{d}{}
}
\description{
森林調査での各調査区-出現全種の行列表
}
\examples{
m<-ForestFloraMatrix(dd5)
head(m)
sp<-names(m)[-1]
df <- data.frame(sp,code_forest=)

ForestFloraCode(cod=m[[c("オオシラビソ")]])

fc<-c(); for( i in sp)fc<-c(fc,ForestFloraCode(cod=m[[i]]))
 forest_code<-data.frame(sp,forest_code=fc)
 # usethis::use_data(forest_code)

}
