% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{ForestTrees}
\alias{ForestTrees}
\title{data/ForestTrees}
\format{
An object of class `data.frame`.
}
\usage{
data(ForestTrees)
}
\description{
data/ForestTrees
}
\examples{
str(ForestTrees)
head(ForestTrees)

}
\keyword{datasets}
