% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{ggplot_contour_data}
\alias{ggplot_contour_data}
\title{Akimaのinterpで出力したコンタデータをggplotのgeom_contour
で描画できるデータフレームに変換する関数}
\usage{
ggplot_contour_data(interp_res = interp_res)
}
\arguments{
\item{interp_res}{Akimaのinterpで出力したリスト型のコンタデータ}
}
\value{
ggplotのgeom_contourで描画できるデータフレーム
}
\description{
Akimaのinterpで出力したコンタデータをggplotのgeom_contour
で描画できるデータフレームに変換する関数
}
\examples{
plot.name<-"Bunazaka" #"Mimatsu"
interp_res <-rasXYZ[[which(names(rasXYZ)==plot.name)]]
interp_df<-ggplot_contour_data(interp_res)
ggplot() +
geom_contour(data = interp_df, aes(x = x, y = y, z = z), color = "blue")

}
