% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{html_family_label}
\alias{html_family_label}
\title{html科名ラベル
html植物目録作成用
種名と科名を列に含んだデータフレームを入力し科名のhtmlラベルを出力する}
\usage{
html_family_label(species_list)
}
\arguments{
\item{species_list}{}
}
\description{
html科名ラベル
html植物目録作成用
種名と科名を列に含んだデータフレームを入力し科名のhtmlラベルを出力する
}
\examples{
spj <- c("ブナ","コナラ","スギ","アカミノイヌツゲ","イタヤカエデ","アオカラムシ","ミズナラ")
species_list <- APG[match(spj,APG$種名),]

html_family_label(species_list)

}
