% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{html_science_name}
\alias{html_science_name}
\title{html書式　属名・種小名をイタリックにする}
\usage{
html_science_name(sn0)
}
\arguments{
\item{sn0}{}
}
\description{
html書式　属名・種小名をイタリックにする
}
\examples{
html_science_name("Quercus serrata Thunb. ex. Muuray")

}
