% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{LandMarks}
\alias{LandMarks}
\title{data/LandMarks　中島さんのデータ}
\format{
An object of class `data.frame`.
}
\usage{
data(LandMarks)
}
\description{
data/LandMarks　中島さんのデータ
}
\examples{
head(LandMarks)
}
\keyword{datasets}
