% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{leg}
\alias{leg}
\title{data/leg　図凡例}
\format{
An object of class `data.frame`.
}
\source{
automatically generated skeleton
}
\usage{
data(leg)
}
\description{
data/leg　図凡例
}
\examples{
leg
}
\keyword{datasets}
