% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs_with_examples.R
\name{package_docs_with_examples}
\alias{package_docs_with_examples}
\title{Generate Package Documentation with Examples and Knit to HTML}
\usage{
package_docs_with_examples(
  package_name,
  output_rmd = "package_docs_with_examples.Rmd",
  output_html = "package_docs_with_examples.html"
)
}
\arguments{
\item{package_name}{The name of the package (character string)}

\item{output_rmd}{The output Rmd filename (default: "package_docs_with_examples.Rmd")}

\item{output_html}{The output HTML filename (default: "package_docs_with_examples.html")}
}
\value{
None. The Rmd and HTML files are generated in the current working directory.
}
\description{
Generate Package Documentation with Examples and Knit to HTML
}
\examples{
package_docs_with_examples("TateyamaVegetation")

}
