% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{plots_5periods_sum}
\alias{plots_5periods_sum}
\title{種の被度の時系列変化　(例　ササの経年変化作図)
Change of coverage for floral types (life form species)}
\usage{
plots_5periods_sum(data = d, condition = form == "sasa")
}
\arguments{
\item{data}{A tibble with columns `form`, `plot`, and `c1:c5`.}
}
\value{
Summarised tibble.
}
\description{
種の被度の時系列変化　(例　ササの経年変化作図)
Change of coverage for floral types (life form species)
}
\examples{
library(magrittr)
library(tidyr)
library(dplyr)
library(ggplot2)

 d<-plots_5periods_sum(data=vc5,condition=form == "sasa")
# c2〜c5をロング化


d_long <- d \%>\%
  select(plot, c2:c5) \%>\%
  pivot_longer(c2:c5, names_to = "period", values_to = "value") \%>\%
  mutate(period = factor(period, levels = c("c2","c3","c4","c5")))

# 各plotを別パネル
ggplot(d_long, aes(x = period, y = value, group = plot)) +
  geom_line() +
  geom_point() +
  facet_wrap(~ plot, scales = "free_y") +
  labs(x = "Period", y = "Coverage", title = "Change in c2–c5 by plot") +
  theme_minimal()


# 1枚で表示
shapes <- c(15, 16, 17, 18, 19, 20, 21, 22, 23)

ggplot(d_long, aes(x = period, y = value,
                   color = plot,
                   shape = plot,
                   group = plot)) +
  geom_line() +
  geom_point(size = 3) +
  scale_shape_manual(values = shapes) +
  labs(x = "Period", y = "Coverage",
       title = "Change in c2–c5 (all plots)") +
  theme_minimal()

#'
}
