% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{plt}
\alias{plt}
\title{data/plt}
\format{
An object of class `data.frame`.
}
\source{
automatically generated skeleton
}
\usage{
data(plt)
}
\description{
data/plt
}
\examples{
plt
}
\keyword{datasets}
