% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{RData_APG}
\alias{RData_APG}
\title{APG植物目録()
米倉浩司・梶田忠 (2007-) 植物和名ー学名インデックスYList
を利用させていただきました}
\usage{
RData_APG()
}
\description{
APG植物目録()
米倉浩司・梶田忠 (2007-) 植物和名ー学名インデックスYList
を利用させていただきました
}
\examples{
head(APG)
sp.<-c("ブナ","ハイマツ","リュウキュウマツ","イソマツ")
cat("以下はAPGリストにありません")
data.frame(sp=sp.[!is.element(sp.,APG$種名)])


}
