% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{RData_flora}
\alias{RData_flora}
\title{立山植生モニタリング調査の植物目録}
\usage{
RData_flora()

RData_flora()
}
\description{
立山植生モニタリング調査の植物目録

立山植生モニタリング調査の植物目録
}
\examples{
head(flora)
flora[flora$美松!="",c("spj","sp")]
sp.<-c("ブナ","ハイマツ","リュウキュウマツ","イソマツ")
cat("以下はfloraリストにありません")
data.frame(sp=sp.[!is.element(sp.,flora$spj)])


head(flora)
flora[flora$美松!="",c("spj","sp")]
sp.<-c("ブナ","ハイマツ","リュウキュウマツ","イソマツ")
cat("以下はfloraリストにありません")
data.frame(sp=sp.[!is.element(sp.,flora$spj)])


}
