% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData_topo.R
\name{RData_rasXYZ}
\alias{RData_rasXYZ}
\title{Title}
\usage{
RData_rasXYZ()
}
\description{
Title
}
\examples{
par(mfrow=c(3,3))
for(i in 1:nrow(plt)){
 contour(rasXYZ[[i]],main=plt$na[i])
}
ggplot() +
geom_contour(data =rasXYZ[[1]], aes(x = x, y = y, z = z), color = "blue")

}
