% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetation_RData.R
\name{RData_subplot_xy}
\alias{RData_subplot_xy}
\title{data/RData　サブプロット}
\usage{
RData_subplot_xy()
}
\description{
data/RData　サブプロット
}
\examples{
.<-subplot_xy
head(.)
table(.$plot)
plot(1:100,type="n")
i<-.$plot=="Bunazaka"
text(.$x[i],.$y[i],.$subplot[i])
RData_subplot_xy()

}
