% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{SpeciesForm}
\alias{SpeciesForm}
\title{Title}
\usage{
SpeciesForm(sp. = c("ブナ", "ハイマツ", "ミヤマカンスゲ", "コケモモ"))
}
\arguments{
\item{sp.}{生活型を知りたい種名のベクトル}
}
\value{
入力した種名の生活型
}
\description{
Title
}
\examples{
SpeciesForm(c("ブナ","ハイマツ","ミヤマカンスゲ","コケモモ"))

}
