% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{SpeciesNameCheck}
\alias{SpeciesNameCheck}
\title{種の和名が植物目録floraに記載されているか確認します}
\usage{
SpeciesNameCheck(spj, spj00 = flora$spj)
}
\arguments{
\item{spj}{種名(和名)のベクトル}

\item{spj00}{参照する種名(和名)のベクトル}
}
\value{
記載の有無、記載のない入力種名
}
\description{
種の和名が植物目録floraに記載されているか確認します
}
\examples{
SpeciesNameCheck(c("ブナ","スギ","クロベ"))
d<-FieldNote_Arimine2019_raw
head(d)
spj<-d$種名
spj
SpeciesNameCheck(spj,spj00=flora$spj)



}
