% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{SpeciesNameCorrect}
\alias{SpeciesNameCorrect}
\title{和名の置換
本調査の種和名はYList(米倉浩司・梶田忠 2007-)の第一和名に基づきます。
同種異名は本コマンドで置換表sp_excに基づき修正できます。}
\usage{
SpeciesNameCorrect(spj)
}
\arguments{
\item{spj}{}
}
\description{
和名の置換
本調査の種和名はYList(米倉浩司・梶田忠 2007-)の第一和名に基づきます。
同種異名は本コマンドで置換表sp_excに基づき修正できます。
}
\examples{
d<-FieldNote_Arimine2019_raw
spj<-d$種名
SpeciesNameCheck(spj)
SpeciesNameCheck(SpeciesNameCorrect(spj))

}
