% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{sptype}
\alias{sptype}
\title{data/sptype コード、生活型、 階層の略号表}
\format{
An object of class `data.frame`.
}
\usage{
data(sptype)
}
\description{
data/sptype コード、生活型、 階層の略号表
}
\examples{
sptype
}
\keyword{datasets}
