% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{subplot_xy}
\alias{subplot_xy}
\title{data/ 各調査地におけるサブプロットの位置座標}
\format{
An object of class `data.frame`.
}
\usage{
data(subplot_xy)
}
\description{
data/ 各調査地におけるサブプロットの位置座標
}
\examples{
subplot_xy
}
\keyword{datasets}
