% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{subplot_xy_ggtext}
\alias{subplot_xy_ggtext}
\title{樹木位置図　樹木位置・ラベル・種名・ランドマーク・コンタ・サブプロット・グリッド}
\usage{
subplot_xy_ggtext(plot.name = "Mimatsu")
}
\arguments{
\item{plot.name}{}
}
\description{
樹木位置図　樹木位置・ラベル・種名・ランドマーク・コンタ・サブプロット・グリッド
}
\examples{
(p<-treemap(plot.name="Mimatsu"))
p+subplot_xy_ggtext("Mimatsu")

}
