% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{TateyamaVegetationPlots}
\alias{TateyamaVegetationPlots}
\title{data/ 立山植生モニタリング調査区一覧表}
\format{
An object of class `spec_tbl_df`.
}
\usage{
data(TateyamaVegetationPlots)
}
\description{
data/ 立山植生モニタリング調査区一覧表
}
\examples{
TateyamaVegetationPlots

}
\keyword{datasets}
