% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{text_species}
\alias{text_species}
\title{各調査地の各調査期における出現種}
\usage{
text_species(plot_name = "Arimine", period = 5, data = vc5)
}
\arguments{
\item{data}{}
}
\description{
各調査地の各調査期における出現種
}
\examples{

text_species(plot_name="Arimine",period=5)


}
