% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{topo_gg}
\alias{topo_gg}
\title{樹木位置図(ggplot　ラベル情報なし)
調査用にはtreemapを利用してください}
\usage{
topo_gg(plot.name = "Mimatsu")
}
\arguments{
\item{plot.name}{}
}
\description{
樹木位置図(ggplot　ラベル情報なし)
調査用にはtreemapを利用してください
}
\examples{
#x11()
pn<-"Mimatsu"
 pn<-"Bijodaira"
(p<-treemap(plot.name=pn))
p+topo_gg(pn)
p+stream1
}
