% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{treemap}
\alias{treemap}
\title{植生調査用樹木位置図作成}
\usage{
treemap(plot.name = "Mimatsu")
}
\arguments{
\item{df}{}
}
\description{
植生調査用樹木位置図作成
}
\examples{

#  x11()
treemap("Bijodaira")
treemap("Bunazaka")
treemap("Bunadaira")
treemap("Kaminokodaira")
treemap("Matsuotoge")
treemap("Kagamiishi")

#'
treemap("Mimatsu")
 treemap("Mimatsu")+ ggplot2::ylim(c(45, 105))


treemap("Arimine")
 treemap("Arimine")+ ggplot2::xlim(c(-5, 55))+ ggplot2::ylim(c(-5, 60))





#'
}
