% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{treesXYZ}
\alias{treesXYZ}
\title{data/ 各調査区樹木位置座標}
\format{
An object of class `data.frame`.
}
\usage{
data(treesXYZ)
}
\description{
data/ 各調査区樹木位置座標
}
\examples{
treesXYZ
}
\keyword{datasets}
