% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vc5}
\alias{vc5}
\title{data/ Vegetation chronosequence summary (data frame)}
\format{
A  `tbl_df`.
}
\usage{
data(vc5)
}
\description{
data/ Vegetation chronosequence summary (data frame)
}
\examples{
VCrepo
vc5
}
\keyword{datasets}
