% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{VegetationFloraCode}
\alias{VegetationFloraCode}
\title{植物目録の各調査区における植生被度分布コードの作成}
\usage{
VegetationFloraCode(cod = 1:10)
}
\arguments{
\item{cod}{}
}
\description{
植物目録の各調査区における植生被度分布コードの作成
}
\examples{

VegetationFloraCode(1:10)

}
