% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{VegetationFloraMatrix}
\alias{VegetationFloraMatrix}
\title{植生調査　調査区-出現種　出現頻度行列の作成
VCrepoより集計　10はAに変換}
\usage{
VegetationFloraMatrix(d = VCrepo)
}
\arguments{
\item{d}{}
}
\description{
植生調査　調査区-出現種　出現頻度行列の作成
VCrepoより集計　10はAに変換
}
\examples{

m<-VegetationFloraMatrix(VCrepo)
head(m)
sp<-names(m)[-1]
vc<-c(); for( i in sp)vc<-c(vc,VegetationFloraCode(cod=m[[i]]))
 vegetation_code<-data.frame(sp,vegetation_code=vc)
 # usethis::use_data(vegetation_code)
}
