% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VegetationSurveyYears}
\alias{VegetationSurveyYears}
\alias{VegetationYearsMatrix}
\title{data/ 植生調査年(スプレッドシート列ラベル)}
\format{
An object of class `tbl_df`.

An object of class `tbl_df`.
}
\source{
automatically generated skeleton
}
\usage{
data(VegetationSurveyYears)

data(VegetationSurveyYears)
}
\description{
data/ 植生調査年(スプレッドシート列ラベル)

data/ 植生調査年
}
\examples{
VegetationSurveyYears
VegetationYearsMatrix
VegetationYearsMatrix[1:5,"Arimine"]
}
\keyword{datasets}
