% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vv}
\alias{vv}
\title{data/ 全期調査区サブプロット植生調査生データ
5期末においてvv5(1-5期)と同じもの}
\format{
An object of class `tbl_df`.
}
\usage{
data(vv)
}
\description{
data/ 全期調査区サブプロット植生調査生データ
5期末においてvv5(1-5期)と同じもの
}
\examples{
vv
# 調査区ごとの抽出
subset(vv,plot=="Arimine")
#listで分割
VV<-split(vv, vv5$plot)
#再結合
dplyr::bind_rows(VV, .id = "plot")

}
\keyword{datasets}
