% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vv4}
\alias{vv4}
\title{data/ 1-4期調査区サブプロット植生調査生データ}
\format{
An object of class `data.frame`.
}
\usage{
data(vv4)
}
\description{
data/ 1-4期調査区サブプロット植生調査生データ
}
\examples{
vv4

}
\keyword{datasets}
