% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{FieldNote_Arimine2019_raw}
\alias{FieldNote_Arimine2019_raw}
\title{FieldNote_Arimine2019_raw}
\format{
An object of class `tbl_df`.
}
\source{
automatically generated skeleton
}
\usage{
data(FieldNote_Arimine2019_raw)
}
\description{
FieldNote_Arimine2019_raw
}
\keyword{datasets}
