% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{FieldNote_FormatArange}
\alias{FieldNote_FormatArange}
\title{スプレッドシートで入力した野帳データ・フレーム(tibble)を
　列名の変更、項目追加し計算用に構成します}
\usage{
FieldNote_FormatArange(d = d, calc = CalculationInfo(2025, "有峰"))
}
\arguments{
\item{d}{野帳データ修正版(tibble)}
}
\description{
スプレッドシートで入力した野帳データ・フレーム(tibble)を
　列名の変更、項目追加し計算用に構成します
}
\examples{


d<- FieldNote_Arimine2019_raw
d                                                                #元の野帳データ
FieldNote_FormatArange(d,calc= CalculationInfo(2025,"有峰"))     #計算用データ


}
