% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{Fig_DominanceValue}
\alias{Fig_DominanceValue}
\title{報告書用作図　総合優占度経年変化
Figure of chronosequence for dominant variance}
\usage{
Fig_DominanceValue(plot_name = "Joudosan", vc_repo = VCrepo)
}
\arguments{
\item{plot_name}{}
}
\description{
報告書用作図　総合優占度経年変化
Figure of chronosequence for dominant variance
}
\examples{
library(dplyr)
library(ggrepel)

Fig_DominanceValue("Kaminokodaira")
Fig_DominanceValue("Arimine")
Fig_DominanceValue("Matsuotoge")

Fig_DominanceValue(plt$plot_name[9])

par(mfrow=c(3,4))
for(i in plt$plot_name)Fig_DominanceValue(i)

par(mfrow=c(1,1))

# 単年度版
Fig_DominanceValue(plot_name="Arimine",vc_repo=VCrepo)

}
