% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{Fig_FrequencyCoverage}
\alias{Fig_FrequencyCoverage}
\title{報告書用作図　出現頻度−平均被度}
\usage{
Fig_FrequencyCoverage(
  plot_name = "Arimine",
  x = "f5",
  y = "c5",
  sp = "sp",
  data = VCrepo
)
}
\arguments{
\item{x}{column name of frequency}

\item{y}{column name of coverage}

\item{sp}{column name of species}

\item{d}{data frame of tibble   VCrepo}
}
\description{
報告書用作図　出現頻度−平均被度
}
\examples{
library(ggplot2)
Fig_FrequencyCoverage(plot_name="Mimatsu")
Fig_FrequencyCoverage(plot_name="Arimine")
Fig_FrequencyCoverage(plot_name="Arimine",x="f5",y="c5")



}
