% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{Fig_FrequencyDominace}
\alias{Fig_FrequencyDominace}
\title{報告書用作図　出現頻度−総合優占度}
\usage{
Fig_FrequencyDominace(
  plot_name = "Arimine",
  x = "f5",
  y = "di5",
  data = VCrepo,
  sp = "sp"
)
}
\arguments{
\item{x}{column name of frequency}

\item{y}{column name of coverage}

\item{data}{list of data frame (tibble)   default :   VCrepo}

\item{sp}{column name of species}
}
\description{
報告書用作図　出現頻度−総合優占度
}
\examples{
　library(ggplot2)
　library(ggrepel)
　
Fig_FrequencyDominace(plot_name="Mimatsu")
Fig_FrequencyDominace(plot_name="Arimine")
Fig_FrequencyDominace(plot_name="Arimine",x="f3",y="di3")
Fig_FrequencyDominace(plot_name="Arimine",x="f5",y="di5")

}
