% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{Fig_FrequencyDominace}
\alias{Fig_FrequencyDominace}
\title{報告書用作図　出現頻度−総合優占度}
\usage{
Fig_FrequencyDominace(
  plot_name = "Arimine",
  x = "f5",
  y = "di5",
  data = VCrepo,
  sp = "sp",
  xlabel = "出現頻度(％)",
  ylabel = "総合優占度"
)
}
\arguments{
\item{plot_name}{name of survey plot  default}

\item{x}{vector of column names of x axis (mean of columes) such as x="f5"  or paste0("f",1:4)}

\item{y}{vector of column names of x axis (mean of columes) such as y="di5" or  paste0("di",1:4)))}

\item{data}{data=VCrepo}

\item{sp}{a column name of species}

\item{xlabel}{text : default is "出現頻度(％)"}

\item{ylabel}{text : default is"総合優占度"}
}
\description{
報告書用作図　出現頻度−総合優占度
}
