% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{FloraListMaker}
\alias{FloraListMaker}
\title{html植物目録の作成}
\usage{
FloraListMaker(spj = sp0, memo = NA)
}
\arguments{
\item{spj}{和名}

\item{memo}{メモ}
}
\value{
htmlリスト
}
\description{
# 米倉浩司・梶田忠 (2003-) 「BG Plants 和名－学名インデックス」（YList），http://ylist.info」
}
\examples{
spj <- c("ブナ","コナラ","スギ","アカミノイヌツゲ","イタヤカエデ","アオカラムシ","ミズナラ")
(fl.<-  FloraListMaker(spj))
memo.<-c("位山","平井","平井","平井","キャンパス","位山","名古屋","位山")
(fl.<-  FloraListMaker(spj,memo.))

# 植生調査植物目録
sp0<-unique(vv$sp)#'
(fl. <- FloraListMaker(sp0))

# cat(fl.,file="data-raw/test_FloraList.html")  # html保存

# 植生調査植物目録(美松・有峰)
sp0<-unique(subset(vv,plot=="Mimatsu")$sp)
# cat(FloraListMaker(sp0),file="data-raw/FloraList_Mimatsu.html")

sp0<-unique(subset(vv,plot=="Arimine")$sp)
# cat(FloraListMaker(sp0),file="data-raw/FloraList_Arimine.html")

 #期末植物目録
  d<-flora2
　memo<-paste("【森】",d$forest_code,"【植】",d$vevetation_code,"【外】",d$NonNative_code)
　str(memo)
　#cat(FloraListMaker(flora2$spj,memo=memo),file="FloraList_V.html")
　
}
