% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flora_list.R
\name{FloraListMaker}
\alias{FloraListMaker}
\title{html植物目録の作成}
\usage{
FloraListMaker(spj, memo)
}
\arguments{
\item{spj}{和名}

\item{memo}{メモ}
}
\value{
htmlリスト
}
\description{
html植物目録の作成
}
\examples{
spj <- c("ブナ","コナラ","スギ","アカミノイヌツゲ","イタヤカエデ","アオカラムシ","ミズナラ")
memo<-c("位山","平井","平井","平井","キャンパス","位山","名古屋","位山")
filename="flora_list_test.htm"

(FloraList.<- FloraListMaker(spj,memo))

#美松調査区の植物目録
names(vv)
names(APG)
spj<-unique(subset(vv,plot=="Mimatsu")$sp)[1:106] #野帳から出現種を抽出

spj[is.na(match(spj,APG$種名))]

memo.<-rep("Mimatsu",length(spj))
(FloraList.<- FloraListMaker(spj,memo=memo.))

#　htmlに保存　ブラウザ用
#cat(FloraList., file=filename)

}
