% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegetationSurvey.R
\name{PlotSpecies_1to4}
\alias{PlotSpecies_1to4}
\title{各調査区の1〜4期の出現種と各期被度合計
※　1期については　Domin_Krajina}
\usage{
PlotSpecies_1to4(site = "Mimatsu")
}
\arguments{
\item{site}{ex.　"Arimine"}
}
\value{
data.frame
}
\description{
各調査区の1〜4期の出現種と各期被度合計
※　1期については　Domin_Krajina
}
\examples{

( df <-　PlotSpecies_1to4 ("Mimatsu"))
 #

 ( df2 <-　PlotSpecies_1to4 ("Arimine"))
 # write.csv(df2, "data_raw/PlotSpecies有峰.csv")

}
