% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegetationSurvey.R
\name{PlotSpecies_1to4}
\alias{PlotSpecies_1to4}
\title{各調査区の1〜4期の出現種と各期被度合計
※　1期については　Domin_Krajina}
\usage{
PlotSpecies_1to4(調査地 = "有峰")
}
\arguments{
\item{調査地}{漢字で入力のこと}
}
\value{
data.frame
}
\description{
各調査区の1〜4期の出現種と各期被度合計
※　1期については　Domin_Krajina
}
\examples{

( df <-　PlotSpecies_1to4 ("美松"))
 # write.csv(df, "data_raw/PlotSpecies美松.csv")

 PlotSpecies_1to4 ("有峰")
 # write.csv(df, "data_raw/PlotSpecies美松.csv")

}
