% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{rasXYZ}
\alias{rasXYZ}
\title{data/各調査区等高線用ラスタデータrasXYZ}
\format{
An object of class `list`.
}
\usage{
data(rasXYZ)
}
\description{
data/各調査区等高線用ラスタデータrasXYZ
}
\examples{
par(mfrow=c(3,3))
for(i in 1:nrow(plt)){
 contour(rasXYZ[[i]],main=plt$na[i])
}



}
\keyword{datasets}
