% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RData.R
\name{RData_subplot_xy}
\alias{RData_subplot_xy}
\title{data/RData}
\usage{
RData_subplot_xy()
}
\description{
data/RData
}
\examples{
.<-subplot_xy
head(.)
table(.$plot)
plot(1:100,type="n")
i<-.$plot=="Bunazaka"
text(.$x[i],.$y[i],.$subplot[i])
RData_subplot_xy()

}
