% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{SpeciesNameCorrect}
\alias{SpeciesNameCorrect}
\title{Title}
\usage{
SpeciesNameCorrect(spj)
}
\arguments{
\item{spj}{}
}
\description{
Title
}
\examples{
d<-FieldNote_Arimine2019_raw
spj<-d$種名
SpeciesNameCheck(spj)
SpeciesNameCheck(SpeciesNameCorrect(spj))

}
