% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{subplot_xy}
\alias{subplot_xy}
\title{subplot_xy}
\format{
An object of class `data.frame`.
}
\source{
automatically generated skeleton
}
\usage{
data(subplot_xy)
}
\description{
subplot_xy
}
\keyword{datasets}
