% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{subplot_xy_ggtext}
\alias{subplot_xy_ggtext}
\title{Title}
\usage{
subplot_xy_ggtext(plot.name = "Mimatsu")
}
\arguments{
\item{plot.name}{}
}
\description{
Title
}
\examples{
(p<-treemap(plot.name="Mimatsu"))
p+subplot_xy_ggtext("Mimatsu")

}
