% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{subplot_xy_text}
\alias{subplot_xy_text}
\title{Title}
\usage{
subplot_xy_text(plot.name = "Mimatsu", ...)
}
\arguments{
\item{plot.name}{}
}
\description{
Title
}
\examples{
plot(1:100,type="n")
subplot_xy_text(plot.name="Mimatsu",col="red",cex=3)
subplot_xy_text(plot.name="Arimine",col="Blue",cex=2)
subplot_xy_text(plot.name="Kaminokodaira",col="Magenta",cex=2)

}
