% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{TateyamaVegetationPlots}
\alias{TateyamaVegetationPlots}
\title{TateyamaVegetationPlots}
\format{
An object of class `spec_tbl_df`.
}
\source{
automatically generated skeleton
}
\usage{
data(TateyamaVegetationPlots)
}
\description{
TateyamaVegetationPlots
}
\keyword{datasets}
