% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{topo_gg}
\alias{topo_gg}
\title{Title}
\usage{
topo_gg(plot.name = "Mimatsu")
}
\arguments{
\item{plot.name}{}
}
\description{
Title
}
\examples{
x11()
pn<-"Mimatsu"
 pn<-"Bijodaira"
(p<-treemap(plot.name=pn))
p+topo_gg(pn)
p+stream1
}
