% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{treemap}
\alias{treemap}
\title{Title}
\usage{
treemap(plot.name = "Mimatsu")
}
\arguments{
\item{df}{}
}
\description{
Title
}
\examples{

#  x11()
treemap("Bijodaira")
treemap("Bunazaka")
treemap("Bunadaira")
treemap("Kaminokodaira")
treemap("Matsuotoge")
treemap("Kagamiishi")

#'
treemap("Mimatsu")
 treemap("Mimatsu")+ ylim(c(45, 105))


treemap("Arimine")
 treemap("Arimine")+ xlim(c(-5, 55))+ ylim(c(-5, 60))





#'
}
