% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{treemap_ggplot}
\alias{treemap_ggplot}
\title{Title}
\usage{
treemap_ggplot(df)
}
\arguments{
\item{df}{}
}
\description{
Title
}
\examples{

df <- data.frame(
x = runif(30, 0, 100),
y = runif(30, 0, 100),
z = runif(30, 0, 100),
dbh = runif(30, 10, 50),
id = paste0(100:129),
species = sample(c("A1", "B2", "C3"), 30, replace = TRUE),
vigor = sample(1:5, 30, replace = TRUE)
)
treemap_ggplot(df)
#'
}
