% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{treesXYZ}
\alias{treesXYZ}
\title{treesXYZ}
\format{
An object of class `data.frame`.
}
\source{
automatically generated skeleton
}
\usage{
data(treesXYZ)
}
\description{
treesXYZ
}
\keyword{datasets}
