% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VC}
\alias{VC}
\title{data/ Vegetation chronosequence}
\format{
An object of class `tbl_df`.
}
\usage{
data(VC)
}
\description{
data/ Vegetation chronosequence
}
\examples{
VC
(d<- VC[["Arimine"]])
nm<-names(d)
 gsub("Mean","m",nm)
 sptype
 d$sp
 d.<-cbind(
 d[,c("sp","Frequencydk01","Frequencyc02","Frequencyc03","Frequencyc04","Frequencyc05")],
 d[,c("FreqCodedk01","FreqCodec02","FreqCodec03","FreqCodec04","FreqCodec05")],
 d[,c("Meandk01","Meanc02","Meanc03","Meanc04","Meanc05")],
 d[,c("DominatIndexdk01dk","DominatIndexc02dk","DominatIndexc03dk","DominatIndexc04dk","DominatIndexc05dk")]
 )
 names(d.)<- c("sp",paste0("f",1:5),paste0("F",1:5),paste0("c",1:5),paste0("di",1:5))

 tibble(d.)
  d. \%>\%
    mutate()

 d \%>\% select(sp, starts_with("Meanc"))

}
\seealso{
[VegetationChronologyTable()]
}
\keyword{datasets}
