% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VCrepo}
\alias{VCrepo}
\title{data/ Vegetation chronosequence summary}
\format{
A list of the objects of class `tbl_df`.
}
\usage{
data(VCrepo)
}
\description{
data/ Vegetation chronosequence summary
}
\examples{
VCrepo
}
\seealso{
[VegetationChronologyTable()]
}
\keyword{datasets}
