% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vegetation_code}
\alias{vegetation_code}
\title{data/ 植物目録各調査区頻度分布コード
植生モニタリング調査(小島覚・石田仁)}
\format{
An object of class `tbl_df`.
}
\usage{
data(vegetation_code)
}
\description{
data/ 植物目録各調査区頻度分布コード
植生モニタリング調査(小島覚・石田仁)
}
\examples{

vegetation_code
 SpeciesNameCheck(vegetation_code$sp)

}
\keyword{datasets}
