% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{VegetationChronologyTable}
\alias{VegetationChronologyTable}
\title{Title}
\usage{
VegetationChronologyTable(VT = VT, VTdk = VTdk)
}
\arguments{
\item{VTdk}{}
}
\description{
Title
}
\examples{

VC<-VegetationChronologyTable(VT=VT,VTdk=VTdk)
# usethis::use_data(VC)
d$Arimine


}
