% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{VegetationChronologyTable_report}
\alias{VegetationChronologyTable_report}
\title{報告書用植生経年変化表}
\usage{
VegetationChronologyTable_report(plot_name = "Arimine", vc = VC)
}
\arguments{
\item{plot_name}{}
}
\description{
報告書用植生経年変化表
}
\examples{
VegetationChronologyTable_report("Bijodaira")
VegetationChronologyTable_report("Arimine")
VCrepo<-lapply(plt$plot_name,VegetationChronologyTable_report)
names(VCrepo)<-plt$plot_name

VCrepo[["Kagamiishi"]]
VCrepo[["Arimine"]]

# usethis::use_data(VCrepo,overwrite = TRUE)

# 報告書貼り付け用スクリプト

　d<-as.data.frame(VCrepo[["Mimatsu"]])
 j<-paste0("f",1:5)  #頻度列
　d[,j]<-100*d[,j]    #　\%　
　d$cal[d$cal==1]<-"*"
 d$cal[d$cal==0]<-""
 d
 # library(clipr)
 # write_clip(d)


}
