% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TateyamaVegetationAnalysis.R
\name{VegetationChronologyTable_report}
\alias{VegetationChronologyTable_report}
\title{Title}
\usage{
VegetationChronologyTable_report(plot_name = "Arimine")
}
\arguments{
\item{plot_name}{}
}
\description{
Title
}
\examples{

VegetationChronologyTable_report("Arimine")
VCrepo<-lapply(plt$plot_name,VegetationChronologyTable_report)
names(VCrepo)<-plt$plot_name

VCrepo[["Kagamiishi"]]

# usethis::use_data(VCrepo,overwrite = TRUE)
}
