% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VegetationSurveyYears}
\alias{VegetationSurveyYears}
\alias{VegetationYearsMatrix}
\title{VegetationSurveyYears}
\format{
An object of class `tbl_df`.

An object of class `tbl_df`.
}
\source{
automatically generated skeleton

automatically generated skeleton
}
\usage{
data(VegetationSurveyYears)

data(VegetationSurveyYears)
}
\description{
VegetationSurveyYears

VegetationYearsMatrix
}
\examples{
VegetationYearsMatrix
VegetationYearsMatrix[1:5,"Arimine"]
}
\keyword{datasets}
