% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VT}
\alias{VT}
\title{VT  Vegtaion cross table for the period V (2020-2025)}
\format{
An object of class `list` of  `tbl_df`.
}
\source{
automatically generated skeleton
}
\usage{
data(VT)
}
\description{
VT  Vegtaion cross table for the period V (2020-2025)
}
\examples{
names(VT)
 VT[["Arimine_c04"]]
 VTdk[["Arimine_c04"]]
}
\seealso{
[VTdk]
}
\keyword{datasets}
