% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vv5}
\alias{vv5}
\alias{dd5}
\alias{VTdk}
\title{data/ 1-5期調査区サブプロット植生調査生データ
vv5 for the period I-V (2020-2025)}
\format{
An object of class `tbl_df`.

An object of class data.frame

An object of class `tbl_df`.
}
\source{
中島・石田　「森林調査」
}
\usage{
data(vv5)

data(vv5)

data(vv5)
}
\description{
data/ 1-5期調査区サブプロット植生調査生データ
vv5 for the period I-V (2020-2025)

data/ 1-5期調査区サブプロット森林調査生データ
dd5 for the period I-V (2020-2025)

data/Domin-Kradina集計表
 VTdk Vegetation cross table (Domin-Kradina)
}
\examples{
# NOTE:
# vv5 is stored in long format.
# For manual editing, it is often convenient to split by plot,
# edit each subset, and then re-bind to vv5.
vv5
#listで分割
(VV5<-split(vv5, vv5$plot))
#再結合
dplyr::bind_rows(VV5, .id = "plot")

# 調査区ごとの抽出
(arimine <- subset(vv5, plot == "Arimine"))


head(dd5)

names(VT5)
 VT[["Arimine_c04"]]
 VTdk[["Arimine_c05"]]
}
\keyword{datasets}
