% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{VV5}
\alias{VV5}
\title{data/ 1-5期調査区サブプロット植生調査生データ(リスト版)
VV5 for the period I-V (2020-2025)}
\format{
An object of class `tbl_df`.
}
\source{
automatically generated skeleton
}
\usage{
data(VV5)
}
\description{
data/ 1-5期調査区サブプロット植生調査生データ(リスト版)
VV5 for the period I-V (2020-2025)
}
\examples{
VV5
#tibbleに結合
bind_rows(VV5,.id="plot")

}
\keyword{datasets}
