% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\docType{data}
\name{vv5}
\alias{vv5}
\alias{dd5}
\alias{VTdk}
\title{data/ 1-5期調査区サブプロット植生調査生データ
vv5 for the period I-V (2020-2025)}
\format{
An object of class `tbl_df`.

An object of class data.frame

An object of class `tbl_df`.
}
\source{
automatically generated skeleton

中島・石田　「森林調査」

automatically generated skeleton
}
\usage{
data(vv5)

data(vv5)

data(vv5)
}
\description{
data/ 1-5期調査区サブプロット植生調査生データ
vv5 for the period I-V (2020-2025)

data/ 1-5期調査区サブプロット森林調査生データ
dd5 for the period I-V (2020-2025)

data/Domin-Kradina集計表
 VTdk Vegetation cross table (Domin-Kradina)
}
\examples{
vv5

head(dd5)

names(VT5)
 VT[["Arimine_c04"]]
 VTdk[["Arimine_c04"]]
}
\keyword{datasets}
