% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegetationSurvey.R
\name{vv_sp_check}
\alias{vv_sp_check}
\title{特定種の野帳記録}
\usage{
vv_sp_check(sp. = "ヒロバスゲ")
}
\arguments{
\item{sp.}{}
}
\value{
データフレーム　特定種の野帳記録
}
\description{
特定種の野帳記録
}
\examples{
vv_sp_check("ヒロバスゲ")

vv_sp_check("ウダイカンバ")

vv_sp_check("ミヤマイタチシダ")



}
